requirements.txt
flask
numpy
pandas
azure-storage-blob
scikit-learn